Work around 'as' attribute being no longer valid for <link /> (#4)
The validation on XML documents has become stricter in Magento 2.2
so we can't use the 'as' attribute for preloading assets. Instead
add a template file where we can do whatever we want.

Potential future work on this is to make the Meanbee_CSSPreload
extension more flexible and support this type of declaration.